Refactor of Vector2 and Vector2Int
- Separate Vector2Int into its own header file.
- Move operator overloads out of the class where appropriate so they are normal functions.
- Replace << friend functions with normal functions.
- Fix all warnings across project.

+++ game.cpp
@@ -124,7 +124,7 @@ void Game::Render(SDL_Renderer* renderer) const {
 
 	// Bottom row.
 	TextRendering::DrawTextAt(renderer, m_player.GetAmmo(), { 30, bottomRowY });
-	TextRendering::DrawTextAt(renderer, m_roundTimer.GetTimeRemaining(), { 118, bottomRowY});
+	TextRendering::DrawTextAt(renderer, static_cast<int>(m_roundTimer.GetTimeRemaining()), { 118, bottomRowY});
 	TextRendering::DrawTextAt(renderer, m_player.GetScore(), { 170, bottomRowY });
 	TextRendering::DrawTextAt(renderer, m_player.GetCombo(), { 192, bottomRowY });
